+++ PSFramework/internal/scripts/strings.ps1
@@ -1,4 +1,3 @@
-﻿Import-LocalizedString -Path (Resolve-Path "$script:ModuleRoot\en-us\stringsAssembly.psd1") -Module PSFramework -Language 'en-US'
-
+﻿Import-LocalizedString -Path "$script:ModuleRoot\en-us\*.psd1" -Module PSFramework -Language 'en-US'
 
 $script:strings = Get-PSFLocalizedString -Module PSFramework